Move CLI clases to cli/ subfolder
