+++ exe/centra_orders
@@ -7,7 +7,7 @@ require "bundler/setup" if ENV['CENTRA_GEM_DEV']
 require "csv"
 
 require "centra"
-require "centra/order_cli"
+require "centra/cli/order_cli"
 
 args = Centra::OrderCLI.parse!(name: 'centra_orders')
