+++ lib/centra/cli/cli_utils.rb
